refactor(hub): move with-shape/money → shape/via-money (grammar correction, no behavior change) (#623)
