+++ packages/hub/__tests__/computed/hardening.test.ts
@@ -9,7 +9,7 @@ import { describe, it, expect } from 'vitest'
 import { z } from 'zod'
 import { createNoydb, MoneyPrecisionError } from '../../src/index.js'
 import { evalComputedFields } from '../../src/with-formula/computed/index.js'
-import { money } from '../../src/with-shape/money/descriptor.js'
+import { money } from '../../src/shape/via-money/descriptor.js'
 import type { NoydbStore, EncryptedEnvelope } from '../../src/kernel/types.js'
 
 function memory(): NoydbStore {
