+++ packages/hub/__tests__/describe-contract.test.ts
@@ -15,7 +15,7 @@
  */
 import { describe, it, expect } from 'vitest'
 import { createNoydb } from '../src/kernel/noydb.js'
-import { money } from '../src/with-shape/money/descriptor.js'
+import { money } from '../src/shape/via-money/descriptor.js'
 import { staticDict } from '../src/with-shape/i18n/dictionary.js'
 import { i18nText } from '../src/with-shape/i18n/core.js'
 import { ref } from '../src/kernel/refs.js'
