+++ packages/hub/__tests__/materialized-views/union-money.test.ts
@@ -3,7 +3,7 @@ import { z } from 'zod'
 import { createNoydb, withMaterializedView } from '../../src/index.js'
 import { withAggregate } from '../../src/with-lookup/aggregate/index.js'
 import { sum, min, max } from '../../src/with-lookup/aggregate/reducers.js'
-import { money } from '../../src/with-shape/money/descriptor.js'
+import { money } from '../../src/shape/via-money/descriptor.js'
 import type { NoydbStore, EncryptedEnvelope } from '../../src/kernel/types.js'
 
 function memory(): NoydbStore {
