+++ packages/hub/__tests__/money/aggregate.test.ts
@@ -3,7 +3,7 @@ import { z } from 'zod'
 import { createNoydb } from '../../src/index.js'
 import { withAggregate } from '../../src/with-lookup/aggregate/index.js'
 import { sum, min, max, count, avg } from '../../src/with-lookup/aggregate/reducers.js'
-import { money, MoneyUnsupportedError } from '../../src/with-shape/money/descriptor.js'
+import { money, MoneyUnsupportedError } from '../../src/shape/via-money/descriptor.js'
 import type { NoydbStore, EncryptedEnvelope } from '../../src/kernel/types.js'
 
 function memory(): NoydbStore {
