+++ packages/hub/__tests__/money/descriptor.test.ts
@@ -5,7 +5,7 @@ import {
   MoneyCurrencyError,
   MoneyPrecisionError,
   MoneyUnsupportedError,
-} from '../../src/with-shape/money/descriptor.js'
+} from '../../src/shape/via-money/descriptor.js'
 import { NoydbError } from '../../src/kernel/errors.js'
 
 describe('money()', () => {
