+++ packages/hub/__tests__/money/nested-paths.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest'
 import { z } from 'zod'
 import { createNoydb, money, ValidationError } from '../../src/index.js'
-import { quantizeMoneyFields, decodeMoneyFields } from '../../src/with-shape/money/normalize.js'
-import { money as moneyFactory } from '../../src/with-shape/money/descriptor.js'
+import { quantizeMoneyFields, decodeMoneyFields } from '../../src/shape/via-money/normalize.js'
+import { money as moneyFactory } from '../../src/shape/via-money/descriptor.js'
 import type { NoydbStore, EncryptedEnvelope } from '../../src/kernel/types.js'
 
 // #334 — nested-path money declarations. The descriptor key is a path:
