+++ packages/hub/__tests__/money/normalize.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
-import { quantizeMoneyFields, decodeMoneyFields } from '../../src/with-shape/money/normalize.js'
-import { money, MoneyPrecisionError } from '../../src/with-shape/money/descriptor.js'
+import { quantizeMoneyFields, decodeMoneyFields } from '../../src/shape/via-money/normalize.js'
+import { money, MoneyPrecisionError } from '../../src/shape/via-money/descriptor.js'
 
 const fields = {
   fixed: { total: money({ currency: 'EUR', scale: 2 }) },
