+++ packages/hub/src/kernel/collection.ts
@@ -11,7 +11,7 @@ import type { I18nTextDescriptor } from '../with-shape/i18n/core.js'
 import { getAtPath, setAtPathInPlace, stripI18nFilled } from '../with-shape/i18n/core.js'
 import type { DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../with-shape/i18n/dictionary.js'
 import { isStaticDictDescriptor } from '../with-shape/i18n/dictionary.js'
-import type { MoneyDescriptor } from '../with-shape/money/descriptor.js'
+import type { MoneyDescriptor } from '../shape/via-money/descriptor.js'
 import { moneyRuntime } from './money-runtime.js'
 import type { ComputedFields } from '../with-formula/computed/index.js'
 import { enforceClassifiedWrite } from '../with-shape/classified/write.js'
