+++ packages/hub/src/kernel/query/scan-builder.ts
@@ -69,7 +69,7 @@ import type {
 } from '../../with-lookup/aggregate/aggregation.js'
 import type { JoinContext, JoinLeg, JoinableSource } from './join.js'
 import { DanglingReferenceError } from '../errors.js'
-import type { MoneyDescriptor } from '../../with-shape/money/descriptor.js'
+import type { MoneyDescriptor } from '../../shape/via-money/descriptor.js'
 import { moneyRuntime } from '../money-runtime.js'
 
 /**
