+++ packages/hub/src/shape/via-money/arith.ts
