+++ packages/hub/src/shape/via-money/descriptor.ts
