+++ packages/hub/src/shape/via-money/paths.ts
